use gateway/api/utils for parsing utilities
